requirements.txt
Django>=3.0,<=3.1
whitenoise
psycopg2-binary
Pillow
mysqlclient
django-crispy-forms
django-markdownx
django-allauth
azure-storage-blob
web3
django-bootstrap-modal-forms
